feat: extract shared public reader package
