+++ apps/felicia-public-site/package.json
@@ -14,6 +14,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@felicia/shared": "workspace:*",
     "maplibre-gl": "^5.9.0",
     "svelte": "^5.38.0"
   },
