+++ apps/felicia-public-site/src/api/adapt.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "bun:test"
-import type { ApiJourney, ApiMemento } from "./types"
+import type { ApiJourney, ApiMemento } from "@felicia/shared"
 import { adaptJourney } from "./adapt"
 import { loadGoldenRouteFixture } from "./fixtures"
