+++ apps/felicia-public-site/src/main.ts
@@ -1,5 +1,5 @@
 import "maplibre-gl/dist/maplibre-gl.css"
-import "./index.css"
+import "@felicia/shared/public.css"
 import App from "./App.svelte"
 import { mount } from "svelte"
