+++ packages/felicia-shared/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@felicia/shared",
+  "private": true,
+  "version": "0.1.0",
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "svelte": "./src/index.ts",
+      "import": "./src/index.ts"
+    },
+    "./public.css": "./src/public.css",
+    "./contracts/public": "./src/contracts/public.ts"
+  },
+  "dependencies": {
+    "maplibre-gl": "^5.9.0",
+    "svelte": "^5.38.0"
+  }
+}
