+++ packages/felicia-shared/src/i18n/catalog.ts
