+++ packages/felicia-shared/src/v4/V4Detail.svelte
