+++ packages/felicia-shared/src/v4/stubs.ts
